requirements.txt
torch
transformers
Pillow
youtube_transcript_api
pytube
yt_dlp
opencv-python
pandas
matplotlib
scikit-learn
